+++ src/acetime/acetz.py
@@ -16,7 +16,6 @@ class acetz(tzinfo):
     """
 
     def __init__(self, zone_info: ZoneInfo):
-        self.zone_info = zone_info
         self.zp = ZoneProcessor(zone_info, use_python_transition=True)
 
     def utcoffset(self, dt: Optional[datetime]) -> timedelta:
@@ -43,6 +42,10 @@ class acetz(tzinfo):
         return timedelta(seconds=offset_info.dst_offset)
 
     def tzname(self, dt: Optional[datetime]) -> str:
+        """Return the abbreviation of the timezone, instead of the full name,
+        for compatibility with other Python timezone libraries (pytz, dateutils,
+        and zoneinfo). Use tzfullname() to get the full name of the time zone.
+        """
         assert dt
         offset_info = self.zp.get_timezone_info_for_datetime(dt)
         if not offset_info:
@@ -87,6 +90,16 @@ class acetz(tzinfo):
         newdt = newutcdt.replace(tzinfo=self, fold=offset_info.fold)
 
         return newdt
+
+    def tzfullname(self, follow_link: bool = False) -> str:
+        """Return the full name of the time zone. If the timezone is a Link and
+        'follow_link' is True, then this returns the name of the target zone.
+        This method does *not* override a method in pareent tzinfo class.
+        """
+        return self.zp.get_name(follow_link)
+
+    def islink(self) -> bool:
+        return self.zp.is_link()
 
     def zone_processor(self) -> ZoneProcessor:
         return self.zp