+++ tests/test_acetz.py
@@ -7,6 +7,7 @@ from acetime.common import SECONDS_SINCE_UNIX_EPOCH
 from acetime.acetz import acetz, ZoneManager
 from acetime.zonedb.zone_registry import ZONE_REGISTRY
 from acetime.zonedb.zone_infos import ZONE_INFO_America_Los_Angeles
+from acetime.zonedb.zone_infos import ZONE_INFO_US_Pacific
 
 
 # Enable logging during unittests.
@@ -68,7 +69,9 @@ class TestLosAngeles(unittest.TestCase):
         self.assertEqual(-8 * 3600, dtc_utcoffset.total_seconds())
 
         assert(dtc.tzinfo is not None)
-        self.assertEqual("PST", dtc.tzinfo.tzname(dtc))
+        self.assertEqual("PST", tz.tzname(dtc))
+        self.assertEqual("America/Los_Angeles", tz.tzfullname())
+        self.assertEqual("America/Los_Angeles", tz.tzfullname(follow_link=True))
 
     def test_fromtimestamp(self) -> None:
         """Create date from AceTime epoch seconds using fromtimestamp()."""
@@ -86,7 +89,9 @@ class TestLosAngeles(unittest.TestCase):
         self.assertEqual(-8 * 3600, dte_utcoffset.total_seconds())
 
         assert(dte.tzinfo is not None)
-        self.assertEqual("PST", dte.tzinfo.tzname(dte))
+        self.assertEqual("PST", tz.tzname(dte))
+        self.assertEqual("America/Los_Angeles", tz.tzfullname())
+        self.assertEqual("America/Los_Angeles", tz.tzfullname(follow_link=True))
 
     def test_before_spring_forward(self) -> None:
         tz = zone_manager.gettz('America/Los_Angeles')
@@ -295,6 +300,7 @@ class TestLosAngeles(unittest.TestCase):
         instead of going through the ZoneManager.
         """
         tz = acetz(ZONE_INFO_America_Los_Angeles)
+        self.assertFalse(tz.islink())
 
         epoch_seconds = 7984800
         unix_seconds = epoch_seconds + SECONDS_SINCE_UNIX_EPOCH
@@ -313,6 +319,40 @@ class TestLosAngeles(unittest.TestCase):
         self.assertEqual(timedelta(hours=1), dtc.dst())
 
         self.assertEqual(dtu, dtc)
+
+
+class TestUSPacific(unittest.TestCase):
+
+    def test_zone_info(self) -> None:
+        """Test creation of acetz object using the US/Pacific ZoneInfo database
+        entry, instead of going through the ZoneManager. This should follow the
+        'link_to' entry in ZoneInfo and use the America/Los_Angeles ZoneEras.
+        """
+        tz = acetz(ZONE_INFO_US_Pacific)
+        self.assertTrue(tz.islink())
+
+        epoch_seconds = 7984800
+        unix_seconds = epoch_seconds + SECONDS_SINCE_UNIX_EPOCH
+        dtu = datetime.fromtimestamp(unix_seconds, tz=timezone.utc)
+
+        dtc = datetime(2000, 4, 2, 3, 0, 0, tzinfo=tz)
+        self.assertEqual(unix_seconds, int(dtc.timestamp()))
+        self.assertEqual(2000, dtc.year)
+        self.assertEqual(4, dtc.month)
+        self.assertEqual(2, dtc.day)
+        self.assertEqual(3, dtc.hour)
+        self.assertEqual(0, dtc.minute)
+        self.assertEqual(0, dtc.second)
+        self.assertEqual("PDT", dtc.tzname())
+        self.assertEqual(timedelta(hours=-7), dtc.utcoffset())
+        self.assertEqual(timedelta(hours=1), dtc.dst())
+
+        self.assertEqual(dtu, dtc)
+
+        assert(dtc.tzinfo is not None)
+        self.assertEqual("PDT", tz.tzname(dtc))
+        self.assertEqual("US/Pacific", tz.tzfullname())
+        self.assertEqual("America/Los_Angeles", tz.tzfullname(follow_link=True))
 
 
 class TestTunis(unittest.TestCase):